fix: resolve remaining compilation errors

+++ app/src/main/java/com/exory550/exoryfilemanager/helpers/crypto/FileCrypto.kt
@@ -8,6 +8,8 @@ import java.security.KeyStore
 import java.security.SecureRandom
 import java.util.*
 import javax.crypto.Cipher
+import javax.crypto.CipherInputStream
+import javax.crypto.CipherOutputStream
 import javax.crypto.KeyGenerator
 import javax.crypto.SecretKey
 import javax.crypto.spec.GCMParameterSpec
@@ -196,7 +198,7 @@ class FileCrypto(private val context: Context) {
                         val zeros = ByteArray(BUFFER_SIZE)
                         file.outputStream().use { out ->
                             for (i in 0 until length step BUFFER_SIZE) {
-                                out.write(zeros, 0, minOf(BUFFER_SIZE, (length - i).toInt()))
+                                out.write(zeros, 0, minOf(BUFFER_SIZE.toLong(), length - i).toInt())
                             }
                         }
                     }
@@ -204,7 +206,7 @@ class FileCrypto(private val context: Context) {
                         val ones = ByteArray(BUFFER_SIZE).apply { fill(0xFF.toByte()) }
                         file.outputStream().use { out ->
                             for (i in 0 until length step BUFFER_SIZE) {
-                                out.write(ones, 0, minOf(BUFFER_SIZE, (length - i).toInt()))
+                                out.write(ones, 0, minOf(BUFFER_SIZE.toLong(), length - i).toInt())
                             }
                         }
                     }
@@ -213,7 +215,7 @@ class FileCrypto(private val context: Context) {
                         file.outputStream().use { out ->
                             for (i in 0 until length step BUFFER_SIZE) {
                                 random.nextBytes(randomBytes)
-                                out.write(randomBytes, 0, minOf(BUFFER_SIZE, (length - i).toInt()))
+                                out.write(randomBytes, 0, minOf(BUFFER_SIZE.toLong(), length - i).toInt())
                             }
                         }
                     }
